unit tests from lists.rs moved to tests/unit/lists.rs
